LayoutTests/editing/style/exec-command-foreColor-with-color-filter-expected.txt


"world" should be #ecfffd:
| <font>
|   color="#ecfffd"
|   "<#selection-anchor>hello world<#selection-focus>"
